+++ src/main/java/dk/kudishin/statsbot/storage/Storage.java
@@ -1,7 +1,7 @@
 package dk.kudishin.statsbot.storage;
 
 import org.springframework.stereotype.Component;
-import dk.kudishin.statsbot.common.BotUser;
+import dk.kudishin.statsbot.data.BotUser;
 import org.telegram.telegrambots.meta.api.objects.Message;
 
 import java.util.ArrayList;
@@ -10,11 +10,11 @@ import java.util.HashSet;
 @Component
 public interface Storage {
 
-    void saveId(Long chatId);
+    void saveId(Integer chatId);
 
-    HashSet<Long> getChatIds();
+    HashSet<Integer> getChatIds();
 
-    void removeId(Long chatId);
+    void removeId(Integer chatId);
 
     void saveMessage(Message message);
